machine-learning/introduction-to-machine-llearning/images/requirements.txt
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
autopep8==1.6.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
blis==0.7.7
brotlipy==0.7.0
catalogue==2.0.7
certifi==2022.5.18.1
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.13.0
conda-package-handling @ file:///C:/ci/conda-package-handling_1649106011304/work
cryptography @ file:///C:/ci/cryptography_1652083563162/work
cycler==0.11.0
cymem==2.0.6
debugpy @ file:///C:/ci/debugpy_1637091961445/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
editdistpy==0.1.3
entrypoints @ file:///C:/ci/entrypoints_1649926621128/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
filelock==3.7.1
flake8==4.0.1
fonttools==4.31.2
huggingface-hub==0.7.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.16.1
ipykernel @ file:///C:/ci/ipykernel_1646982785443/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1651601994004/work
jedi @ file:///C:/ci/jedi_1644315428289/work
Jinja2==3.1.2
joblib==1.1.0
jupyter-client @ file:///C:/ci/jupyter_client_1650515989971/work
jupyter-core @ file:///C:/ci/jupyter_core_1651656286743/work
kiwisolver==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
menuinst @ file:///C:/ci/menuinst_1631733438520/work
mpmath==1.2.1
murmurhash==1.0.7
nest-asyncio @ file:///C:/ci/nest-asyncio_1649829929390/work
networkx==2.7.1
nltk==3.7
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathy==0.6.1
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.0
preshed==3.0.6
prompt-toolkit==3.0.29
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycodestyle==2.8.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.8.2
pyflakes==2.4.0
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.7
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2022.1
PyWavelets==1.3.0
pywin32==302
PyYAML==6.0
pyzmq @ file:///C:/ci/pyzmq_1638435148211/work
regex==2022.6.2
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
scikit-image==0.19.2
scikit-learn==1.1.2
scipy==1.8.0
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==5.2.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
sympy==1.10.1
symspellpy==6.7.6
textblob==0.17.1
thinc==8.0.17
threadpoolctl==3.1.0
tifffile==2022.3.25
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado @ file:///C:/ci/tornado_1606924294691/work
tqdm @ file:///C:/ci/tqdm_1650636210717/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
transformers==4.19.2
typer==0.4.1
typing_extensions==4.1.1
urllib3 @ file:///C:/ci/urllib3_1650639883891/work
wasabi==0.9.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
